updated style.css
